Migrate PluginsDslIT tests to new dsl
^KT-45745 In Progress
