+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/kotlin/org/jetbrains/kotlin/gradle/BaseGradleIT.kt
@@ -29,6 +29,7 @@ import org.junit.runner.RunWith
 import java.io.File
 import java.nio.file.Files
 import java.util.regex.Pattern
+import javax.xml.parsers.DocumentBuilderFactory
 import kotlin.io.path.isDirectory
 import kotlin.test.*
 
@@ -1008,3 +1009,73 @@ Finished executing task ':$taskName'|
 fun BaseGradleIT.BuildOptions.withFreeCommandLineArgument(argument: String) = copy(
     freeCommandLineArgs = freeCommandLineArgs + argument
 )
+
+private const val MAVEN_LOCAL_URL_PLACEHOLDER = "<mavenLocalUrl>"
+internal const val PLUGIN_MARKER_VERSION_PLACEHOLDER = "<pluginMarkerVersion>"
+
+internal fun BaseGradleIT.transformProjectWithPluginsDsl(
+    projectName: String,
+    wrapperVersion: GradleVersionRequired = defaultGradleVersion,
+    directoryPrefix: String? = null,
+    minLogLevel: LogLevel = LogLevel.DEBUG
+): BaseGradleIT.Project {
+
+    val result = Project(projectName, wrapperVersion, directoryPrefix, minLogLevel)
+    result.setupWorkingDir()
+
+    val settingsGradle = File(result.projectDir, "settings.gradle").takeIf(File::exists)
+    settingsGradle?.modify {
+        it.replace(MAVEN_LOCAL_URL_PLACEHOLDER, MavenLocalUrlProvider.mavenLocalUrl)
+    }
+
+    result.projectDir.walkTopDown()
+        .filter {
+            it.isFile && (it.name == "build.gradle" || it.name == "build.gradle.kts" ||
+                    it.name == "settings.gradle" || it.name == "settings.gradle.kts")
+        }
+        .forEach { buildGradle ->
+            buildGradle.modify(::transformBuildScriptWithPluginsDsl)
+        }
+
+    return result
+}
+
+internal fun transformBuildScriptWithPluginsDsl(buildScriptContent: String): String =
+    buildScriptContent.replace(PLUGIN_MARKER_VERSION_PLACEHOLDER, KOTLIN_VERSION)
+
+/** Copies the logic of Gradle [`mavenLocal()`](https://docs.gradle.org/3.4.1/dsl/org.gradle.api.artifacts.dsl.RepositoryHandler.html#org.gradle.api.artifacts.dsl.RepositoryHandler:mavenLocal())
+ */
+private object MavenLocalUrlProvider {
+    /** The URL that points to the Gradle's mavenLocal() repository. */
+    val mavenLocalUrl by lazy {
+        val path = propertyMavenLocalRepoPath ?: homeSettingsLocalRepoPath ?: m2HomeSettingsLocalRepoPath ?: defaultM2RepoPath
+        File(path).toURI().toString()
+    }
+
+    private val homeDir get() = File(System.getProperty("user.home"))
+
+    private fun getLocalRepositoryFromXml(file: File): String? {
+        if (!file.isFile)
+            return null
+
+        val xml = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(file)
+        val localRepoNodes = xml.getElementsByTagName("localRepository")
+
+        if (localRepoNodes.length == 0)
+            return null
+
+        val content = localRepoNodes.item(0).textContent
+
+        return content.replace("\\$\\{(.*?)\\}".toRegex()) { System.getProperty(it.groupValues[1]) ?: it.value }
+    }
+
+    private val propertyMavenLocalRepoPath get() = System.getProperty("maven.repo.local")
+
+    private val homeSettingsLocalRepoPath
+        get() = getLocalRepositoryFromXml(File(homeDir, ".m2/settings.xml"))
+
+    private val m2HomeSettingsLocalRepoPath
+        get() = System.getProperty("M2_HOME")?.let { getLocalRepositoryFromXml(File(it, "conf/settings.xml")) }
+
+    private val defaultM2RepoPath get() = File(homeDir, ".m2/repository").absolutePath
+}
