--- libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/pluginsDsl/applyAllPlugins/settings.gradle
@@ -1,10 +1,0 @@
-pluginManagement {
-    repositories {
-        maven { url '<mavenLocalUrl>' }
-        gradlePluginPortal()
-    }
-
-    plugins {
-        id "org.jetbrains.kotlin.test.fixes.android" version "$test_fixes_version"
-    }
-}
