+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/pluginsDsl/applyToSubprojects/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'org.jetbrains.kotlin.jvm' version '<pluginMarkerVersion>' apply false
+    id 'org.jetbrains.kotlin.jvm' apply false
 }
 
 allprojects {
