+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/pluginsDsl/applyToSubprojects/settings.gradle
@@ -1,12 +1,1 @@
-pluginManagement {
-    repositories {
-        maven { url '<mavenLocalUrl>' }
-        gradlePluginPortal()
-    }
-
-    plugins {
-        id "org.jetbrains.kotlin.test.fixes.android" version "$test_fixes_version"
-    }
-}
-
 include 'subproject'
